2026.01.19

+++ Mini_Project/src/main/java/com/Library_Management_System/Mini_Project/Repository/MemberRepository.java
@@ -3,5 +3,6 @@ package com.Library_Management_System.Mini_Project.Repository;
 import com.Library_Management_System.Mini_Project.Entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<Member, Integer> {
+public interface MemberRepository extends JpaRepository<Member, Long>{
+
 }
